# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Task API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Test Task API Host"
        }
    ],
    "paths": {
...
        "/item/{public_id}": {
            "get": {
                "tags": [
                    "item"
                ],
                "summary": "View the item",
                "operationId": "Api\\V1\\Http\\Controllers\\ItemsController::view",
                "parameters": [
                    {
                        "name": "public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "item"
                ],
                "summary": "Delete the item",
                "operationId": "Api\\V1\\Http\\Controllers\\ItemsController::delete",
                "parameters": [
                    {
                        "name": "public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "item"
                ],
                "summary": "Update the item",
                "operationId": "Api\\V1\\Http\\Controllers\\ItemsController::update",
                "parameters": [
                    {
                        "name": "public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/item": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "Create the item",
                "operationId": "Api\\V1\\Http\\Controllers\\ItemsController::create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "description": "The link to new resource (/api/v1/item/{public_id})",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Item": {
                "properties": {
                    "public_id": {
                        "type": "string",
                        "example": "cd86bc2"
                    },
                    "name": {
                        "type": "string",
                        "example": "Name"
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    }
                },
                "type": "object"
            },
            "ValidationErrors": {
                "properties": {
                    "errors": {
                        "properties": {
                            "attribute": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "http": {
                "type": "http",
                "description": "YWRtaW46YWRtaW4= -> base64_encode('admin:admin')",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "http",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "description": "JWT.access_token",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}